Move bridge classes
